More loader/source refactoring. Remove AbstractSpatialDataSource, make TableSource parent of Shapes/Point source classes
