+++ packages/file-types/spatial-zarr/src/SpatialDataShapesSource.js
@@ -4,7 +4,7 @@
 /* eslint-disable no-undef */
 import WKB from 'ol/format/WKB.js';
 import { basename } from '@vitessce/zarr';
-import AbstractSpatialDataSource from './AbstractSpatialDataSource.js';
+import SpatialDataTableSource from './SpatialDataTableSource.js';
 
 /** @import { DataSourceParams } from '@vitessce/types' */
 /** @import { TypedArray as ZarrTypedArray, Chunk } from 'zarrita' */
@@ -82,7 +82,7 @@ function toFloat32Array(input) {
   throw new TypeError('Input must be Float32Array, Float64Array, or BigInt64Array');
 }
 
-export default class SpatialDataShapesSource extends AbstractSpatialDataSource {
+export default class SpatialDataShapesSource extends SpatialDataTableSource {
   /**
    *
    * @param {string} path A path to within shapes.
@@ -105,38 +105,6 @@ export default class SpatialDataShapesSource extends AbstractSpatialDataSource {
   }
 
   /**
-   * Class method for loading the obs index.
-   * @param {string|undefined} path
-   * @param {string|undefined} tablePath
-   * @returns {Promise<string[]>} An promise for a zarr array containing the indices.
-   */
-  async loadObsIndex(path = undefined, tablePath = undefined) {
-    // TODO: if a tablePath is provided, use it to load the obsIndex.
-    // Otherwise use the index column from the parquet table.
-
-    let indexPath = getIndexPath(path);
-    if (tablePath) {
-      // TODO: given a path to the shapes,
-      // is there a better way to know which table annotates it
-      // (without the manually-specified table path)?
-      // Reference: https://github.com/scverse/spatialdata/issues/298#issuecomment-1718161329
-      const obsPath = `${tablePath}/obs`;
-      const { _index } = await this.getJson(`${obsPath}/.zattrs`);
-      indexPath = `${obsPath}/${_index}`;
-      const {
-        instance_key: instanceKey,
-        // TODO: filter table index by region and element type.
-        // region_key: regionKey,
-        // region,
-      } = await this.loadSpatialDataElementAttrs(tablePath);
-
-      indexPath = `${obsPath}/${instanceKey}`;
-    }
-    // TODO: support loading from parquet if no tablePath was provided.
-    return this._loadColumn(indexPath);
-  }
-
-  /**
    * Class method for loading general numeric arrays.
    * @param {string} path A string like obsm.X_pca.
    * @returns {Promise<Chunk<any>>} A promise for a zarr array containing the data.
@@ -155,8 +123,10 @@ export default class SpatialDataShapesSource extends AbstractSpatialDataSource {
       };
     }
     const parquetPath = getParquetPath(path);
-    const arrowTable = await this.loadParquetTable(parquetPath);
-    const columnArr = arrowTable.getChild(basename(path))?.toArray();
+    const columnName = basename(path);
+    const columns = [columnName];
+    const arrowTable = await this.loadParquetTable(parquetPath, columns);
+    const columnArr = arrowTable.getChild(columnName)?.toArray();
     return {
       shape: [columnArr.length],
       // TODO: support other kinds of TypedArrays via @vitessce/arrow-utils.
@@ -242,6 +212,26 @@ export default class SpatialDataShapesSource extends AbstractSpatialDataSource {
         return coords[0];
       },
     );
+  }
+
+  /**
+   * 
+   * @param {string} elementPath 
+   * @returns {Promise<Array<any>|null>}
+   */
+  async loadShapesIndex(elementPath) {
+    const formatVersion = await this.getShapesFormatVersion(elementPath);
+    if (formatVersion === '0.1') {
+      // Shapes v0.1 did not use Parquet, so we use the parent Zarr-based column loading function.
+      return this._loadColumn(getIndexPath(elementPath));
+    }
+
+    const parquetPath = getParquetPath(elementPath);
+    const indexColumn = await this.loadParquetTableIndex(parquetPath);
+    if (indexColumn) {
+      return indexColumn.toArray();
+    }
+    return null;
   }
 
   /**
