+++ packages/file-types/spatial-zarr/src/spatialdata-loaders/SpatialDataObsPointsLoader.js
@@ -98,14 +98,17 @@ export default class SpatialDataObsPointsLoader extends AbstractTwoStepLoader {
   }
 
   async loadObsIndex() {
-    // TODO: remove this function and just use the one from the dataSource?
-    // But why is the index column name always different?...
-
-    const { path } = this.options;
-    // TODO: will the label column of the parquet table always be numeric?
-    const arr = await this.dataSource.loadNumeric(getIndexPath(path));
-    const obsIds = Array.from(arr.data).map(i => String(i));
-    return obsIds;
+    const { tablePath, path } = this.options;
+    if (tablePath) {
+      return this.dataSource.loadObsIndex(tablePath);
+    }
+    const indexColumn = await this.dataSource.loadPointsIndex(path);
+    if (indexColumn) {
+      const obsIds = Array.from(indexColumn).map(i => String(i));
+      return obsIds;
+    }
+    // TODO: if still no index column (neither from AnnData.obs.index nor from parquet table index),
+    // then create an index based on the row count?
   }
 
   async load() {
